somthing/requirements.txt
anyio==4.0.0
asttokens==2.3.0
backcall==0.2.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
decorator==5.1.1
elevenlabs==0.2.24
emoji==2.8.0
exceptiongroup==1.1.3
executing==1.2.0
gpt4all==1.0.9
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
ipython==8.15.0
jedi==0.19.0
matplotlib-inline==0.1.6
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.39
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.12
pydub==0.25.1
Pygments==2.16.1
requests==2.31.0
romajitable==0.0.1
six==1.16.0
sniffio==1.3.0
sounddevice==0.4.6
srt==3.5.3
stack-data==0.6.2
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.7.1
urllib3==2.0.4
voicevox-client==0.2.2
vosk==0.3.45
wcwidth==0.2.6
websockets==11.0.3